Added mainwindow.ui
